requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
cachetools==6.2.6
certifi==2026.1.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
fastapi==0.128.0
grpcio==1.76.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
numpy==2.4.1
openai==2.16.0
orjson==3.11.6
pandas==3.0.0
protobuf==6.33.4
pydantic==2.12.5
pydantic_core==2.41.5
pymilvus==2.6.8
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
setuptools==80.10.2
six==1.17.0
sniffio==1.3.1
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uvicorn==0.40.0
